revert long->int

+++ src/gt4py/next/program_processors/codegens/gtfn/itir_to_gtfn_ir.py
@@ -54,7 +54,7 @@ def pytype_to_cpptype(t: str):
             "float": "double",
             "float32": "float",
             "float64": "double",
-            "int": "int",
+            "int": "long",
             "int32": "std::int32_t",
             "int64": "std::int64_t",
             "bool": "bool",
